leaudio: Fix possible double MTU request
It might happen, that profile is notified twice about encryption beeing completed.
One from  the callback set in BTM_SetEncryption and second from the GATT
Callback.

It does not happen always, but if happen it  could lead to situation that in the
OnEncryptionComplete callback code is doing double MTU request and CCC
registration.

This patch fixes that.

Bug: 255930205
Test: atest BluetoothInstrumentationTests
Test: manual multiple reconnection of Le Audio device.
Tag: #feature
Change-Id: Ib947e1dbd24a819a011a38c16c7ddde1dca8bc18