Defer disconnect commands until post connecting state
If a device receives a disconnect event while connecting, that event is
currently dropped and the device won't end up disconnecting as a result
of the request.

To fix this, we'll defer the request until we exit the connecting state.
Either we end up disconnecting (connection fails) and the disconnect is
redundant, or we end up in connected where the request is properly
processed

Tag: #stability
Bug: 298209581
Test: atest BluetoothInstrumentationTests
Google: 2736114
Change-Id: I66bc568b0649dd8259d3cdfa1a7979c53851ba12